Split HandJointsManager into two
