+++ BasicSample/Assets/HandTracking/Scripts/FeatureUsageHandJointsManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 936d2f565a176044893c82d55d36a6c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
