+++ BasicSample/Assets/HandTracking/Scripts/OpenXRExtensionHandJointsManager.cs.meta
